Moved BlockChain<T>.MineBlock() tests to a new partial class
